feat: 퀴즈부분 로직

+++ src/main/java/com/memoring/memoring_server/domain/user/User.java
@@ -43,7 +43,7 @@ public class User extends AuditableEntity {
         return user;
     }
 
-    public void addCoin(int count) {
+    public void addCoin(Long count) {
         this.coin += count;
     }
